telegram-bot/requirements.txt
# Core dependencies
aiogram==3.20.0
aiohttp==3.10.5
asyncpg==0.29.0
python-dotenv==1.0.1

# OpenAI for OCR and GPT
openai==1.30.0

# Image processing
Pillow==10.4.0
pdf2image==1.17.0

# Fuzzy matching
rapidfuzz==3.9.6

# Redis for caching
redis==5.0.8

# Utilities
pydantic==2.8.2
pydantic-settings==2.4.0

# Logging
loguru==0.7.2

# Development
pytest==8.3.2
pytest-asyncio==0.23.8
black==24.8.0
flake8==7.1.1